Conflicts:
	src/CodeGen_LLVM.cpp

+++ src/Generator.cpp
@@ -218,7 +218,7 @@ void GeneratorBase::build_params() {
 std::vector<Internal::Parameter> GeneratorBase::get_filter_parameters() {
     build_params();
     std::vector<Internal::Parameter> result;
-    for (int i = 0; i < filter_arguments.size(); ++i) {
+    for (size_t i = 0; i < filter_arguments.size(); ++i) {
         result.push_back(*filter_params[filter_arguments[i].name]);
     }
     return result;
